Version 3.0: Allow copy/pasting of an item and has a history you can ctrl-u and ctrl-k (up/down)
